+++ model/apop_normal.c
@@ -27,7 +27,7 @@ See also the \ref apop_multivariate_normal.
 
 #include "apop_internal.h"
 
-static double positive_sigma_constraint(apop_data *data, apop_model *v){
+static long double positive_sigma_constraint(apop_data *data, apop_model *v){
     //constraint is 0 < beta_2
     Staticdef(apop_data *, constraint, apop_data_falloc((1,1,2), 0, 0, 1));
     return apop_linear_constraint(v->parameters->vector, constraint, 1e-5);
